+++ src/components/Player.vue
@@ -128,11 +128,11 @@
               <font-awesome-icon icon="hand-point-right" />
               Nomination
             </li>
-            <li @click="movePlayer()">
+            <li @click="movePlayer()" :class="{ disabled: session.lockedVote }">
               <font-awesome-icon icon="redo-alt" />
               Move player
             </li>
-            <li @click="swapPlayer()">
+            <li @click="swapPlayer()" :class="{ disabled: session.lockedVote }">
               <font-awesome-icon icon="exchange-alt" />
               Swap seats
             </li>
@@ -143,7 +143,7 @@
               <font-awesome-icon icon="chair" />
               Empty seat
             </li>
-            <li @click="removePlayer">
+            <li @click="removePlayer" :class="{ disabled: session.lockedVote }">
               <font-awesome-icon icon="times-circle" />
               Remove
             </li>
@@ -759,7 +759,8 @@ li.move:not(.from) .player .overlay svg.move {
   }
 
   li.disabled {
-    cursor: default;
+    cursor: not-allowed;
+    opacity: 0.5;
     &:hover {
       color: white;
     }